Moved search to individual calls
